  requirements.txt
flask
llama-index
llama-index-readers-file
python-docx
pydantic
docx2txt